ref(meta): Rename api hooks.
