+++ frontend/src/api/useGetSchools.ts
@@ -1,10 +1,10 @@
-import useApi from "../../api/useApi"
+import useApi from "./useApi"
 
 export interface Schools {
   id: string
   name: string
 }
-export const useQueryAllSchools = (): [Schools[], boolean, () => void] => {
+export const useGetSchools = (): [Schools[], boolean, () => void] => {
   const [schools, loading, setAsOutdated] = useApi<Schools[]>("/user/schools")
   return [schools ?? [], loading, setAsOutdated]
 }
