+++ frontend/src/api/useGetStudents.ts
@@ -1,10 +1,10 @@
-import useApi from "../../api/useApi"
+import useApi from "./useApi"
 
 export interface Student {
   id: string
   name: string
 }
-export const useQueryAllStudents = (
+export const useGetStudents = (
   schoolId: string
 ): [Student[], boolean, () => void] => {
   const [students, isLoading, setAsOutdated] = useApi<Student[]>(
